aelfDevPro/backend/requirements.txt
Flask
openai
google-cloud-storage
